app: 게시판 공개 화면 4개 (W2-13)
P-203~P-206 이 실제 트리에 붙었다. 라우트의 Permission 은 넷 다 비어 있다 —
게시판 권한은 스코프라서 경로만으로 판정할 수 없다. slug 가 게시판을 가리키고
role_permissions 가 가진 것은 board_id 다. 핸들러가 게시판을 읽은 뒤 CanOn 을
부른다. D15 4.2 의 "게이트는 접두사를, 핸들러는 대상을 본다"가 가장 뚜렷한
형태로 나오는 자리다.

권한 없는 게시판은 404 다. 403 은 그 게시판이 있다는 사실을 알려준다 —
/board/staff 가 어디인지 배우는 것과 튕겨 나가는 것의 차이다 (SC-1 4항).
남의 글 수정도 같은 이유로 404 다.

폼은 스키마에서 생성된다. 코드에 필드 목록이 없다 (D14 3절 규칙 1). board_id
와 author_id 는 폼에서 읽지 않는다 (SC-2 5항) — 게시판은 경로에서, 작성자는
세션에서 온다. 보내면 스키마에 없는 키라 422 로 거부된다.

다른 게시판의 글 id 를 공개 게시판 slug 로 열 수 없다. slug 가 권한 판정을
지고 있으므로 짝이 안 맞는 쌍을 받아들이면 공개 URL 로 비공개 글을 읽는다.

조회수는 세션당 한 번 는다. 기억하는 목록에 상한을 뒀다 — 세션 행은 매 요청
기록되고, 크롤러가 여는 글마다 늘어나면 그 자체가 부하다.

변이 둘이 처음에 안 물었다 (비밀글 권한·allow_secret). 저장소 테스트는 있는데
HTTP 층 테스트가 없어서였다 — 핸들러가 저장소에 무엇을 넘기는지는 저장소
테스트가 보지 않는다. 목록·상세·작성 세 경로를 HTTP 로 걷는 테스트를 넣었다.

조회수·합계를 int64 로 바꿨다. D17 의 number 함수가 int64 를 받는데 저장소가
int 를 주고 있었다 — 기존 템플릿 렌더링 테스트가 잡았다.

변이 검증: 11/11.
게이트: make check 0, make test-integration 0.

+++ internal/content/post_integration_test.go
@@ -136,7 +136,7 @@ func TestSecretPostsAreFilteredInSQL(t *testing.T) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		if n != tc.want {
+		if n != int64(tc.want) {
 			t.Errorf("%s: count %d, want %d — 목록과 합계가 다르면 페이저가 거짓말한다", name, n, tc.want)
 		}
 	}
